Fix check backlogged cursors without consumer (#6766)
* fix the backlogged cursors bug

* add some test

* change test case

* change test case

* change test case

* change check logical

* move backlogged cursor check after consumer add
